refactor(app): create dice and pawn services
